requirements.txt
python-dateutil
pandas
requests
beautifulsoup4
